requirements.txt
Django==4.1.5
Jinja2==3.1.2
python-liquid==1.7.0
